+++ src/components/BalancesAccountDetails.tsx
@@ -1,38 +1,43 @@
 import React from 'react'
 import { Link } from 'react-router-dom';
-import { deleteDoc, doc, collection } from 'firebase/firestore';
+import { deleteDoc, doc,  } from 'firebase/firestore';
 import { db, auth } from '../config/firebase'
-import { accountDetailsSlice, selectAccountDetails } from '../redux/accountDetailsSlice';
+import {  selectAccountDetails } from '../redux/accountDetailsSlice';
 import { useSelector } from 'react-redux';
-
+ 
 
 const BalancesAccountDetails = () => {
     const accountDetails = useSelector(selectAccountDetails)
 
-    const deleteCard = async(cardId: string) => {
+    const handleDeleteCard = async (cardId: string | undefined) => {
+        if (!cardId) {
+          console.error('Invalid cardId:', cardId);
+          return;
+        }      
+        try {
 
-        try{
             if(auth.currentUser) {
+                // Delete the document from Firestore
                 await deleteDoc(doc(db, 'accounts', cardId))
                 console.log('Card deleted from Firestore');
-
-                // dispatch(removeCard(cardId))
-            } else {
-                console.log('User not Logged in')
-            }
-        } catch(error) {
-            console.log('Error deleting card', error)
+            
+                } else {
+                   console.log('User not Logged in')
+                }
+          
+        } catch (error: any) {
+          // Log error details
+          console.error('Error deleting card:', error.message, error.code);
         }
-
-    }
+      }; 
 
   return (
-    <div className=''>
-        <h3 className='text-left text-gray02 text-[22px]' onClick={() => console.log(accountDetailsSlice)}>Account Details</h3>
+    <div className='w-full'>
+        <h3 className='text-left text-gray02 text-[22px]' >Account Details</h3>
 
-        <div className='w-[1104px] h-[292px] mt-4 bg-[#FFF] px-8 pt-8 text-left rounded-lg'>
+        <div className=' mt-4 bg-[#FFF] px-8 pt-8 text-left rounded-lg pb-10'>
             {accountDetails && (
-                <div className='grid grid-cols-3 gap-y-10 '>
+                <div className='grid grid-cols-2 lg:grid-cols-3 gap-8 lg:gap-y-10 justify-center items-start'>
                 <div>
                     <p className='text-gray03'>Bank Name</p>
                         <p className='text-lg font-bold text-secondary'>{accountDetails["cardType"]}</p>
@@ -63,7 +68,7 @@ const BalancesAccountDetails = () => {
 
                 <Link to="/Home/Balances">
                 <button className='text-gray01 font-semibold text-sm'
-                onClick={() => deleteCard(accountDetails.cardType)}
+                onClick={() => handleDeleteCard(accountDetails.id)}
                 >Remove</button> 
                 </Link>
